requirements.txt
asgiref==3.8.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.2.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==11.3.0
PyJWT==2.10.1
sql==2022.4.0
SQLAlchemy==2.0.27
sqlparse==0.5.3
typing_extensions==4.10.0
tzdata==2025.2
Werkzeug==3.0.1
